dom/tests/mochitest/bugs/test_bug563487.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=563487
-->
<head>
  <!-- Put this before the mochi* stuff, since that adds various listeners -->
  <script>
    var exception = null;
    try {
      window.removeEventListener("foo", function() {}, false);
    } catch (e) {
      exception = e;
    }
  </script>
  <title>Test for Bug 563487</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=563487">Mozilla Bug 563487</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 563487 **/
is(exception, null, "removeEventListener threw");
</script>
</pre>
</body>
</html>
